refactor linodeIPs.ts
