+++ src/services/linodes/linode.schema.ts
@@ -78,3 +78,11 @@ export const RebuildLinodeSchema = object({
   stackscript_data: array().of(object()),
   booted: boolean().notRequired()
 });
+
+export const IPAllocationSchema = object({
+  type: string()
+    .required('IP address type (IPv4) is required.')
+    .oneOf(['ipv4'], "Only IPv4 addresses can be allocated."),
+  public: boolean()
+    .required("Must specify public or private IP address.")
+})
