+++ src/services/linodes/linodeIPs.ts
@@ -2,6 +2,22 @@ import { API_ROOT } from 'src/constants';
 
 import Request, { setData, setMethod, setURL } from '../index';
 
+import { IPAllocationSchema } from './linode.schema';
+
+type IPAddress = Linode.IPAddress;
+
+export interface IPAllocationRequest {
+  type: 'ipv4';
+  public: boolean;
+}
+
+/**
+ * getLinodeIPs
+ * 
+ * Return a list of IP addresses allocated to this Linode.
+ * 
+ * @param linodeId { number } The id of the Linode whose addresses you would like to retrieve.
+ */
 export const getLinodeIPs = (id: number) =>
   Request<Linode.LinodeIPsResponse>(
     setURL(`${API_ROOT}/linode/instances/${id}/ips`),
@@ -9,18 +25,19 @@ export const getLinodeIPs = (id: number) =>
   )
     .then(response => response.data);
 
-/** @todo type. */
-export const allocatePrivateIP = (linodeID: number) =>
-  Request(
+/**
+ * allocateIPAddress
+ * 
+ * Allocates a public or private IPv4 address to a Linode
+ * 
+ * @param linodeId { number } The id of the Linode to receive a new IP address.
+ * @param data { object }
+ * @param data.type { string } Must be "ipv4", as currently only IPv4 addresses can be allocated.
+ * @param data.public { boolean } True for a public IP address, false for a private address.
+ */
+export const allocateIPAddress = (linodeID: number, data: IPAllocationRequest) =>
+  Request<IPAddress>(
     setURL(`${API_ROOT}/linode/instances/${linodeID}/ips`),
     setMethod('POST'),
-    setData({ type: 'ipv4', public: false }),
-  ).then(response => response.data);
-
-/** @todo type */
-export const allocatePublicIP = (linodeID: number) =>
-  Request(
-    setURL(`${API_ROOT}/linode/instances/${linodeID}/ips`),
-    setMethod('POST'),
-    setData({ type: 'ipv4', public: true }),
+    setData(data, IPAllocationSchema),
   ).then(response => response.data);
