+++ src/utilities/allocateIPAddress.ts
@@ -1,0 +1,7 @@
+import { allocateIPAddress } from 'src/services/linodes';
+
+export const allocatePublicIP = (linodeId: number) =>
+  allocateIPAddress(linodeId, { type: 'ipv4', public: true})
+
+export const allocatePrivateIP = (linodeId: number) =>
+  allocateIPAddress(linodeId, { type: 'ipv4', public: false})
